Create finalProject.cpp
Write a program to keep track of the amount of food that one monkey eats in one day.   The program should have an array of 24 elements and allow the user to input the amount of food in grams that the monkey eats every hour in a single day so values in a (0.00 to 100.00 range).   Once all data has been entered then the program should list the hour that the monkey ate the most plus the average food per hour that the monkey consumed.  Remember to validate the input and put comments into your program as well as name/date and e-mail.